feat: replace text component with rich text & remove children from button component
